feat: implement database relationships and audit fields
- Convert all entity IDs from int to Long for scalability and database best practices
- Use @GeneratedValue(strategy = GenerationType.IDENTITY) for auto-incrementing IDs
- Implement JPA relationships:
  * User (1) ---> (*) Booking: @OneToMany with @JoinColumn
  * Booking (*) ---> (1) User: @ManyToOne with nullable=false
  * Booking (*) ---> (1) Slot: @ManyToOne with nullable=false
  * FitnessCentre (1) ---> (*) Slot: @OneToMany with CascadeType.ALL
  * Slot (*) ---> (1) FitnessCentre: @ManyToOne with nullable=false
- Add JPA audit fields to all entities:
  * createdAt with @CreatedDate
  * updatedAt with @LastModifiedDate
- Create JpaAuditConfig with @EnableJpaAuditing
- Add unique constraints:
  * FitnessCentre.name: UNIQUE
  * Slot.(fitnessCentre, date, startTime): UNIQUE composite
- Add @Column(nullable = false) to all required fields
- Use FetchType.LAZY for all relationships (performance optimization)
- Update repositories to work with relationships
- Remove SequenceGenerator dependency (database handles ID generation)
- Update services to use object references instead of primitive IDs
- Update controllers to use Long IDs in path variables
- Update all DTOs and tests to use new relationship structure

Data Integrity Improvements:
- Foreign key constraints prevent orphaned records
- Unique constraints prevent duplicates
- Cascading delete for Centre -> Slots
- Lazy loading for better query performance
- Automatic audit trail with createdAt/updatedAt

Test Results:
- Total: 136 tests
- Passed: 136
- Coverage: >85%
- All relationships working correctly
- All tests use new entity structure

Database Schema Changes:
- User: AUTO_INCREMENT id (was manual)
- Booking: AUTO_INCREMENT id, new user_id and slot_id FKs
- Slot: AUTO_INCREMENT id, new fitness_centre_id FK
- FitnessCentre: AUTO_INCREMENT id, UNIQUE name constraint
- All tables now have created_at and updated_at audit columns

+++ src/main/java/com/rsh/fitness_centre/controller/BookingController.java
@@ -52,7 +52,7 @@ public class BookingController {
       @ApiResponse(responseCode = "404", description = "Booking not found")
   })
   public Booking cancelBooking(
-      @Parameter(description = "Booking ID") @PathVariable int bookingId) {
+      @Parameter(description = "Booking ID") @PathVariable Long bookingId) {
     return bookingService.cancelBooking(bookingId);
   }
 
@@ -64,7 +64,7 @@ public class BookingController {
       @ApiResponse(responseCode = "404", description = "Booking not found")
   })
   public Booking getBooking(
-      @Parameter(description = "Booking ID") @PathVariable int bookingId) {
+      @Parameter(description = "Booking ID") @PathVariable Long bookingId) {
     return bookingService.getBooking(bookingId);
   }
