+++ src/main/java/com/rsh/fitness_centre/entity/request/AddBookingRequest.java
@@ -12,11 +12,11 @@ public class AddBookingRequest {
   @NotNull(message = "Slot ID is required")
   @Positive(message = "Slot ID must be positive")
   @Schema(description = "ID of the activity slot to book", example = "1")
-  private Integer slotId;
+  private Long slotId;
   
   @NotNull(message = "User ID is required")
   @Positive(message = "User ID must be positive")
   @Schema(description = "ID of the user making the booking", example = "1")
-  private Integer userId;
+  private Long userId;
 
 }
