+++ src/main/java/com/rsh/fitness_centre/repository/FitnessCentreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FitnessCentreRepository extends JpaRepository<FitnessCentre, Integer> {
+public interface FitnessCentreRepository extends JpaRepository<FitnessCentre, Long> {
   
   Optional<FitnessCentre> findByName(String name);
 }
